Documentation and cleanup.
